+++ release/v/vietnam/HISTORY.md
@@ -1,6 +1,13 @@
 vietnam Change History
 ====================
 
+1.0.3 (2025-03-10)
+----------------
+* Updated by cathaylab
+* fixing several bugs in generation
+* added possibility to type Zhuang and Cantonese romanization
+* added more consonants
+
 1.0.2 (2025-03-03)
 ----------------
 * Updated by cathaylab